Fix YEAR workaround for prepared statements
In https://github.com/vitessio/vitess/pull/7020 the logic was changed to
always map YEAR to binary. But that is wrong, since this breaks the text
protocol.

Instead, we need to make sure that when we parse the binary data for
prepared statements, we treat YEAR there as binary. That ensures we can
parse it the way it's encoded in practice in the protocol (and not how
it is documented).

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>