Small internal cleanups
This cleans up some logic around the default collation and coercion. We
now make the MySQL 8 collations the default if there's no version
mismatch instead of what we had with MySQL 5.7.

Additionally we fix a small issue that doesn't really matter in
practice with an unknown collation, but we know have a full type so we
do know it, so let's use it.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>