Add resultsObserver to ScatterConn
Updates ScatterConn's ExecuteMultiShard and StreamExecuteMulti methods
to accept a results observer parameter. These methods call
resultsObserver#observe with every sqltypes.Result object returned from
the underlying shards. This makes it straightforward to add
instrumentation that needs access to the individual results actually
returned from the shards, not just the Result object returned to the
client which may or may not include data from all the shards

This PR adds the resultObserver, but it is set to a nullResultsObserver
so it is effectively a no-op. Having this extensibility will make it
much easier for downstream forks to add instrumentation as needed.

Signed-off-by: Rafer Hazen <rafer@ralua.com>