vtbackup: allow exiting if failing early
The control flow in here relies on panic'ing and recovering from that
panic.

Exiting the process ends up being blocked on this defer block before the
panic recovery handler if we exit before `servenv.ExitChan` is created.
We're then ultimately blocking forever and require a kill -9 while we're
trying to write into a nil channel.

This is ultimately becuase we start up `servenv.RunDefault()` in a
goroutine, and `servenv.ExitChan` ends up being created within that, so
it's possible we are trying to exit before the channel is created.

Signed-off-by: Matt Robenolt <matt@ydekproductions.com>