Handle encoding binary data separately
Binary data should be encoded as such when interpolated, for example in
bind variables. Otherwise valuable type information that the value is
actually binary is lost.

Also means that we should not mark literals in the parser as binary
either. They are strings and there's other introducers like `_binary`
that end up marking things as proper binary.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>