Fix incorrect typing used for prepared statements
This has been using all binary everywhere since day one this was added,
but we are a lot more type aware and handle types correctly now.

This means we also should convert prepared statement values to the
proper types as well when we receive them as inputs for queries. This
way we know we can handle them correctly further down as well.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>