Track type for id column for messages
This tracks the explicit type for the id column so that we can use that
properly when binding variables. If we don't bind the right type, we end
up depending on implicit MySQL type casting behavior which isn't
something we should do.

Making it explicit here is much more robust.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>