Use _binary prefix
Using the X'' syntax makes things behave differently, since it makes
MySQL interpret it as a hex number in contexts where it casts it to an
integer, for example in a `where id = 'binary string'` where `id` is a
numerical type.

`_binary` as a prefix more behaves like no prefix at all, so it better
matches how things behave today so we're better backwards compatible.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>