Deflake vdiffs in e2e tests using load generator
This impacted e2e tests that were not previously using VPlayerBatching.

The load generator constantly generates INSERTs, which are then
effeciently batched in vplayer so we get ~ 7x more throughput than
before and thus we need more time for filtered replication to catch
up after we've stopped it for the vdiff. ESPECIALLY since we're
using the --update-table-stats flag and the ANALYZE TABLE and its
locking causes a pause in updates to the table the load generator is
inserting into -- in particular for the test clusters that only
have PRIMARY tablets as everything is interacting directly.

Signed-off-by: Matt Lord <mattalord@gmail.com>