Update bundled MySQL version to 8.0.40
We haven't updated the bundled MySQL version in a while and it also
blocks us from upgrading our base Debian version for the containers we
build as well.

So update to the latest 8.0.x version here. It's also at least version
8.0.35, which means that updates after this one are going to be much
easier as upgrades / downgrades between different patch versions are
possible for 8.0.35 and later.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>