Ignore execution time errors for schemadiff view analysis
The UnsupportedConstruct error is returned for execution of select
queries inside vtgate, but when we analyze views we don't care about
those.

Views here are executed at the MySQL level though and always inside a
single shard, so we don't want to return these errors.

The test example here shows a case where it thinks it's in sharded mode
because of a missing dependency, but we only want to show that we're
missing a table dependency.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>