Fix vtcombo parsing flags incorrectly
Inside vtcombo we were setting up the vtenv too early in `init()`. This
leads to reading the MySQL version number before it actually is set from
the given flags, ignoring the passed in option.

This moves it into the `run` function which is the appropriate place.

We do a similar fix for `vtctldclient` by making sure we initialize the
vtenv inside the pre-run command instead of in the `init()` to avoid a
potentially similar problem.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>