evalengine: normalize types during compilation
The evalengine compiler does not operate on arbitrary width types: once
a value is evaluated into the compiler stack, its type is fixed to the
widest possible for the representation (e.g. all signed values are
promoted to INT64, all unsigned values to UINT64, etc).

The BindVariable code was not taking this into account when evaluating
and pushing the value of a bind variable into the evaluation stack. A
bind variable with type INT32 was being pushed into the stack as an
INT64, but it kept its original type when statically typing the rest of
the stack, so it caused bad compilation.

Signed-off-by: Vicent Marti <vmg@strn.cat>