Implement server side cursors
To be able to use the MySQL foreign data wrapper for Postgres, we have to
support server side cursors. To that end each connection now keeps track
of a list of open cursors, and implements the COM_STMT_FETCH command for
retrieving data from an open cursor.

Based on the implementation in the Dolthub fork of Vitess:
https://github.com/dolthub/vitess/pull/228
https://github.com/dolthub/vitess/pull/232

Co-Authored-By: James Cor <jcor@ucsd.edu>
Co-Authored-By: Aaron Son <aaron@dolthub.com>
Signed-off-by: Daniël van de Giessen <daniel@dvdgiessen.nl>