Switch back things to Oracle runners
This enables the Oracle runners again, but it doesn't remove the
app_armor cleanup logic. I want to look into how that is failing because
we probably either need to keep this change, or make it more resilient
since we see failures that point at the app armor issue on these
workers.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>