Add forward compatibility for caching_sha2_password and replication
This extracts the logic from #18010 for replication setup. There are
cases during upgrades and downgrades where the following can happen:

- The replication user is created on a newer version with
  caching2_sha_password.
- We downgrade from this newer version to an older version.
- The older version doesn't have the logic to add the
  GET_SOURCE_PUBLIC_KEY flag so it fails to re-enable replication.

What this tells us, is that we need to first ship the change here to
allow for future forward compatibility in v22, so we can change the
default safely in v23.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>