Partially revert the binary encoding logic for safer upgrade
This reverts part of the binary encoding change to keep encoding as
plain text even for binary data, to stay compatible with the previous
version of Vitess.

This applies only to v22, since on main we're on v23 and that is
compatible with the change in v22 already. This is only a concern during
upgrades.

Fixes #18143

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>