Use "status" integer instead of "more" boolean
The previous commit which implemented support for server side cursors
also updated the function signature of the MySQL query functions to
return a integer "status" value instead of a boolean "more" value. One
of the flags in that status integer represents the "more" flag. This
commit updates all places where the MySQL query functions are called to
use this new status integer.

Signed-off-by: Daniël van de Giessen <daniel@dvdgiessen.nl>