Fix for simple projection showing no fields
When a simple projection ends up in a statement that is prepared, we end
up directly calling `buildFields` from `GetFields`.

When we do that, we don't do the same check we do on results, which is
that we have a names only optimization where `sc.Cols` is nil. This
leads to building an empty fields list.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>