Fix regression in v22 around new flag setup
In v22, a new global flag was added in go/mysql which causes issues.
Instead, we need to setup a configuration option like we do for all the
other settings and pass those in from the appropriate places.

I have here explicitly opted for not refactoring the function arguments,
since this is a regression we should backport and I want to do the most
minimal work possible here.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>