discovery: clarify use of TabletFilter in NewHealthCheck
While updating calling code that uses this function, I noticed that the
TabletFilter argument is implied to both be a slice and a required argument
with one or more values.

Since TabletFilter is an interface but the TabletFilters slice type implements
that interface, I believe the confusion stems from there.

Document that passing a type implementing TabletFilter is entirely optional given
that all internal callers also guard against nil with a nil check. Rename the
argument to the singular version to reinforce the notion that it is an interface
type.

Signed-off-by: Matt Layher <mdlayher@planetscale.com>