Add support for querylog-emit-on-any-condition-met flag
This PR implements the RFC as described in https://github.com/vitessio/vitess/issues/18459

Currently there are several querylog filter/threshold flags you can specify (e.g. query log-row-threshold, querylog-filter-tag). Each flag's purpose is clearly defined when it is used standalone. However when you specify a combination of those flags, the result is less clear and hard to reason or remember.

In many situations, the users just want to have an OR relationship between those flags, they want to see vtgate/vttablet emit querylog either when the row count exceeded or when the query contains the specified filter tag.

A new querylog flag --querylog-emit-on-any-condition-met will be added to emit querylog when any of the querylog threshold/filter/mode is met.

As part of the implementation, we also change the return signature of function ShouldEmitLog to return both the boolean decision and a string EmitReason to explain why it should emit the log.  EmitReason is a comma-separated string to contain the conditions which triggers the should decision, sample value: filtertag,row indicates the function calculates that it should emit the log because both filtertag matches the specified querylog-tag and the rows returned/affected in the query exceeding the querylog-row-threshold.  The caller vtgate/logstats and vttablet/tabletserver/tabletenv/logstats are modified to write out the new EmitReason as a new column in the querylog output file.

Refactor the function ShouldEmitLog to deal with the legacy code path as well as the new code path when querylog-emit-on-any-condition-met is specified.  A new function ShouldEmitLogOnCondition is abstracted out to test an individual querylog condition plus doing aggregation of allMatches and reasons array.

Unit tests are added/modified to test the new querylog-emit-on-any-condition flag and the new EmitReason return value

Signed-off-by: Henry Cai <henry.cai@slack-corp.com>