Add new `force` flag to `DemotePrimary` to force a failover during `EmergencyReparentShard`.
During `EmergencyReparentShard` we don't want to wait for the old primary to get acks for all pending changes, because we're performing a demotation at the same time as unlinking existing replicas from the primary. If all replicas are unlinked while commits are still pending, the primary will never be able to receive the semi sync acks to be unblocked, causing ERS to never be able to complete.

Signed-off-by: Arthur Schreiber <arthur@planetscale.com>