Use vreplication transaction lag in VTAdmin
The vreplication lag metrics are liveness metrics and not what a
VTAdmin is interested in when trying to understand if a vreplication
workflow is caught up, lagging, etc. The liveness metric can be 0
while the workflow is really lagging by hours, which is reflected in
the vreplication transaction lag which compares the timestamp of the
transaction from the source with the timestamp for when we executed
it in the vstream.

h/t to Claude Code for doing all of the vtadmin work!

Signed-off-by: Matt Lord <mattalord@gmail.com>