Fix incorrect escaping of dual in canonical formatting
In the canonical formatting mode, we ended up with `dual` being escaped
with backticks. This is wrong, as `dual` with backticks looks for the
actual table named `dual` in MySQL (which would not exist) and not the
virtual dual table.

So the formatting needs to treat the table name special if it's dual.

Signed-off-by: Dirkjan Bussink <d.bussink@gmail.com>