querythrottler: Remove noisy config loading error log
Remove the "Error loading config" log message that fires when the
throttler config file doesn't exist. This log is misleading because:

1. Query Throttler is disabled by default
2. Missing config file is expected behavior (falls back to Enabled=false)
3. This is not an error condition that requires user attention

This file-based config polling mechanism is being replaced with a
topology-server watch in an upcoming re-architecture, at which point
the file loader will be removed entirely.

Relates to: https://github.com/vitessio/vitess/issues/18869

Signed-off-by: siddharth16396 <siddharth16396@gmail.com>