Re-add `options` to `QueryService` interface
https://github.com/vitessio/vitess/pull/19007 updated the `QueryService`
interface with a new `Session` interface that contains the session UUID
and the existing `ExecuteOptions`. Since the `options` parameter was now
redundant and can be pulled from the `Session` directly, it was removed.
This caused a lot of unnecessary changes, but also introduced a race
condition in `scatter_conn.go`:

```diff
	allErrors := stc.multiGoTransaction(
		// ...
		func(rs *srvtopo.ResolvedShard, i int, info *shardActionInfo) (*shardActionInfo, error) {
			// ...

			// Goroutine A reading session
			if session != nil && session.Session != nil {
				opts = session.Options
			}

			// Goroutine B writing session
			if opts == nil && fetchLastInsertID {
				opts = &querypb.ExecuteOptions{FetchLastInsertId: fetchLastInsertID}
+				session = econtext.NewSafeSession(&vtgatepb.Session{Options: opts})
			}

			// ...
		}
	)
```

This PR removes `GetOptions` from the new `Session` interface and adds
back the `options` parameter to the `QueryService` interface. This
allows each goroutine in that `scatter_conn.go` logic to keep its own
local `opts` and not overwrite the shared `session`, reverting the logic
to how it was before.

Signed-off-by: Mohamed Hamza <mhamza@fastmail.com>