Generate race unit tests
Add the race unit tests to the workflow generation logic. This will help
ensure they stay in sync. As a result, this also fixes the issue that
the race unit tests workflow would not fail in CI due to a missing
`set -exo pipefail`, allowing the tests to fail but the overall job to
still succeed.

Signed-off-by: Mohamed Hamza <mhamza@fastmail.com>