Change tracingService methods to use contexts
New and NewFromString in the tracingService interface are called by
trace.NewSpan and trace.NewFromString, who take and return a context.

This aligns them, so the plugin matches their caller.

This makes the interface more compatible with OpenTelemetry API, making it
easier to write an upcoming tracing plugin using that API.

Signed-off-by: Matthew McPherrin <git@mcpherrin.ca>